fix: add strict jinja templating
