+++ docetl/operations/utils/progress.py
@@ -1,0 +1,69 @@
+from typing import Iterable, Optional, Union
+from concurrent.futures import as_completed
+from tqdm import tqdm
+
+class RichLoopBar:
+    """A progress bar class that integrates with Rich console."""
+    
+    def __init__(
+        self,
+        iterable: Optional[Union[Iterable, range]] = None,
+        total: Optional[int] = None,
+        desc: Optional[str] = None,
+        leave: bool = True,
+        console=None,
+    ):
+        if console is None:
+            raise ValueError("Console must be provided")
+        self.console = console
+        self.iterable = iterable
+        self.total = self._get_total(iterable, total)
+        self.description = desc
+        self.leave = leave
+        self.tqdm = None
+
+    def _get_total(self, iterable, total):
+        if total is not None:
+            return total
+        if isinstance(iterable, range):
+            return len(iterable)
+        try:
+            return len(iterable)
+        except TypeError:
+            return None
+
+    def __iter__(self):
+        self.tqdm = tqdm(
+            self.iterable,
+            total=self.total,
+            desc=self.description,
+            file=self.console.file,
+        )
+        for item in self.tqdm:
+            yield item
+
+    def __enter__(self):
+        self.tqdm = tqdm(
+            total=self.total,
+            desc=self.description,
+            leave=self.leave,
+            file=self.console.file,
+        )
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.tqdm.close()
+
+    def update(self, n=1):
+        if self.tqdm:
+            self.tqdm.update(n)
+
+def rich_as_completed(futures, total=None, desc=None, leave=True, console=None):
+    """Yield completed futures with a Rich progress bar."""
+    if console is None:
+        raise ValueError("Console must be provided")
+
+    with RichLoopBar(total=total, desc=desc, leave=leave, console=console) as pbar:
+        for future in as_completed(futures):
+            yield future
+            pbar.update()
